--- src/components/NewRunForm.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default () => {
-  return <div>yo</div>;
-};
