+++ src/pages/index.tsx
@@ -1,17 +1,14 @@
 import React from "react";
-import { useQuery, useMutation } from "@apollo/react-hooks";
+import { useQuery } from "@apollo/react-hooks";
 import gql from "graphql-tag";
 import Container from "@icgc-argo/uikit/Container";
 import { css } from "emotion";
 import Typography from "@icgc-argo/uikit/Typography";
 import RunsTable from "../components/RunsTable";
 import { useAppContext } from "../context/App";
-import Button from "@icgc-argo/uikit/Button";
-import Modal from "@icgc-argo/uikit/Modal";
 import DNALoader from "@icgc-argo/uikit/DnaLoader";
 import { ModalPortal } from "../App";
-import InputLabel from "@icgc-argo/uikit/form/InputLabel";
-import Input from "@icgc-argo/uikit/form/Input";
+import NewRunFormModal from "../components/NewRunFormModal";
 
 export type Run = {
   run_id: string;
@@ -40,57 +37,6 @@ export default () => {
    * modal stuff
    */
   const [loading, setLoading] = React.useState(false);
-
-  const [runWorkflow] = useMutation<
-    any,
-    { workflow_url: string; workflow_params: { [k: string]: any } }
-  >(
-    gql`
-      mutation($workflow_url: String!, $workflow_params: JSON!) {
-        runWorkflow(
-          workflow_url: $workflow_url
-          workflow_params: $workflow_params
-        ) {
-          run_id
-          __typename
-        }
-      }
-    `
-  );
-
-  const [newRunModalShown, setNewRunModalShown] = React.useState(false);
-
-  const onNewRunClick: React.ComponentProps<typeof Button>["onClick"] = () => {
-    setNewRunModalShown(true);
-  };
-
-  const [workflow_url, setWorkflowUrl] = React.useState("");
-  const [workflow_params, setWorkflowParams] = React.useState({});
-
-  const onNewRunConfirmed: React.ComponentProps<
-    typeof Modal
-  >["onActionClick"] = async () => {
-    runWorkflow({
-      variables: {
-        workflow_url: workflow_url,
-        workflow_params: workflow_params
-      }
-    });
-    setLoading(true);
-    await new Promise(resolve => {
-      setTimeout(() => {
-        resolve();
-      }, 1000);
-    });
-    setLoading(false);
-    setNewRunModalShown(false);
-  };
-
-  const onNewRunCanceled: React.ComponentProps<
-    typeof Modal
-  >["onCancelClick"] = () => {
-    setNewRunModalShown(false);
-  };
 
   const { doesPoll } = useAppContext();
 
@@ -162,33 +108,7 @@ export default () => {
           margin: 10px 0px;
         `}
       >
-        <Button onClick={onNewRunClick}>new run</Button>
-        {newRunModalShown && (
-          <ModalPortal>
-            <Modal
-              title="Some famcy form to input params"
-              onCancelClick={onNewRunCanceled}
-              onActionClick={onNewRunConfirmed}
-            >
-              <InputLabel>
-                Workflow URL
-                <Input
-                  aria-label="workflow_url"
-                  value={workflow_url}
-                  onChange={e => setWorkflowUrl(e.target.value)}
-                />
-              </InputLabel>
-              <InputLabel>
-                Workflow Params
-                <Input
-                  aria-label="workflow_params"
-                  value={JSON.stringify(workflow_params)}
-                  onChange={e => setWorkflowParams(JSON.parse(e.target.value))}
-                />
-              </InputLabel>
-            </Modal>
-          </ModalPortal>
-        )}
+        <NewRunFormModal setLoading={setLoading}/>
       </div>
       <Container
         className={css`
